Updated models and connected DbContext
