+++ ProjectAPI/Models/Door.cs
@@ -9,13 +9,11 @@ namespace ProjectAPI.Models
         public int DoorId { get; set; }
 
         [MaxLength(30), Required] //Setting the limit and that it can't be Null (NotNull)
-        public string DoorNumber { get; set; }
+        public string DoorName { get; set; }
 
-        [MaxLength(30), Required]
-        public string DoorLocation { get; set; }
+        // Many-to-many relationship with Tag
+        public ICollection<DoorTag> Tags { get; set; }
 
-        public Code Code { get; set; } //Creating the association
 
-        public int CodeId { get; set; } //Instering the foreign key 
     }
 }
