+++ ProjectAPI/Models/DoorTag.cs
@@ -8,10 +8,13 @@ namespace ProjectAPI.Models
 {
     public class DoorTag //Composite key
     {
-        public Door Door { get; set; }
         public int DoorId { get; set; }
-        public Tag Tag { get; set; }
+
+        public Door Door { get; set; }
+
         public int TagId { get; set; }
+
+        public Tag Tag { get; set; }
     }
 }
